changes eval routines compatible with 2017

+++ vs10/DataAnalysis/@dataset/rateplot.m
@@ -141,7 +141,7 @@ else,
         xlog125;
     end
     try,
-        LegStr = cellify(getGUIdata(gcf,'LegendStr', {}));
+        LegStr = cellify(getGUIdata(double(gcf),'LegendStr', {}));
         LegStr = [LegStr [num2str(max(D.Stim.SPL)) ' dB SPL']];
 %         setGUIdata(gcf,'LegendStr', LegStr);
         legend(LegStr, 'location', 'southwest');
